added tests

+++ src/main/java/ru/yandex/practicum/filmorate/repository/FilmRepository.java
@@ -31,5 +31,5 @@ public interface FilmRepository {
     // Получения списка фильмов по режиссеру
     List<Film> getFilmsByDirector(int directorId, String sortBy);
 
-    List<Film> getCommonFilms(Long userId, Long friendId);
+    List<Film> getCommonFilms(int userId, int friendId);
 }
