+++ src/main/java/ru/yandex/practicum/filmorate/serviceDB/FilmServiceDB.java
@@ -28,5 +28,5 @@ public interface FilmServiceDB {
     // Получения списка фильмов по режиссеру
     List<Film> getFilmsByDirector(int directorId, String sortBy);
 
-    List<Film> getCommonFilms(Long userId, Long friendId);
+    List<Film> getCommonFilms(int userId, int friendId);
 }
